// static/protected/contract.d.ts
// AUTO-GENERATED by scaffold-api-types.ts — DO NOT EDIT
// Generated: 2026-04-05T09:31:24.243Z

// ── src\routes\doc.ts ──

/** GET /docs  ⚠ no schema */
export interface GetDocsRoute {
  path: "/docs";
  method: "get";
  query: any;
  body: any;
  params: any;
  response: unknown;
  errors: never[];
}

// ── src\routes\index.ts ──

/** GET /  ⚠ no schema */
export interface GetRootRoute {
  path: "/";
  method: "get";
  query: any;
  body: any;
  params: any;
  response: { message: string };
  errors: never[];
}

/** GET /health  ⚠ no schema */
export interface GetHealthRoute {
  path: "/health";
  method: "get";
  query: any;
  body: any;
  params: any;
  response:
    | {
        Environment: {
          name: string;
          db: {
            connection: string;
            host: string;
            port: string;
            database: string;
            username: string;
            password: string;
          };
          redis: { host: string; password: string; port: string };
          r2: {
            account_id: string;
            access_key: string;
            secret_access_key: string;
            bucket_name: string;
            endpoint: string;
          };
          socials: { google: { client_id: string; client_secret: string } };
          better_auth_secret: string;
          is_in_development: boolean;
          email: { service: string; user: string; password: string };
        };
      }
    | { message: string };
  errors: never[];
}

// ── src\routes\storage-bucket.ts ──

/** POST /upload-url */
export interface PostUploadUrlRoute {
  path: "/upload-url";
  method: "post";
  query: Record<string, never>;
  body: {
    fileName: string;
    fileCategory: "images" | "pdfs";
    contentType: string;
  };
  params: Record<string, never>;
  response: { key: string; signedUrl: string };
  errors: [{ statusCode: 409; message: "Failed to get signed url" }];
}

// ── src\routes\test.ts ──

/** GET /test */
export interface GetTestRoute {
  path: "/test";
  method: "get";
  query: {
    name?: string;
    age?: number;
  };
  body: Record<string, never>;
  params: Record<string, never>;
  response: { message: string };
  errors: never[];
}

/** GET /test/address */
export interface GetTest_addressRoute {
  path: "/test/address";
  method: "get";
  query: {
    name?: string;
    age?: number;
    address?: string;
  };
  body: Record<string, never>;
  params: Record<string, never>;
  response: { message: string };
  errors: never[];
}

// ── All routes union ──
export type ApiRoutes =
  | GetDocsRoute
  | GetRootRoute
  | GetHealthRoute
  | PostUploadUrlRoute
  | GetTestRoute
  | GetTest_addressRoute;